+++ build.gradle
@@ -1,19 +1,10 @@
 /*
- * build.gradle
  *
- * Copyright (c) 2014-2016, Daniel Ellermann
+ *  * Copyrights 2002-2016 Webb Fontaine
+ *  * Date: 18/08/2017
+ *  * This software is the proprietary information of Webb Fontaine.
+ *  * Its use is subject to License terms.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
  */
 
 
@@ -27,61 +18,45 @@ buildscript {
     }
 }
 
-version '2.0.6'
-group 'com.webbfontaine.grails.plugins'
+allprojects {
+    version '3.0.0'
+    group 'com.webbfontaine.grails.plugins'
 
-apply plugin: 'eclipse'
-apply plugin: 'idea'
-apply plugin: 'org.grails.grails-plugin'
-apply plugin: 'org.grails.grails-plugin-publish'
-apply plugin: 'org.grails.grails-gsp'
-apply plugin: 'org.grails.grails-doc'
-
-repositories {
-    mavenLocal()
-    maven { url 'https://repo.grails.org/grails/core' }
 }
 
-dependencies {
-    compile 'org.springframework.boot:spring-boot-autoconfigure'
-    compile 'org.springframework.boot:spring-boot-starter-actuator'
-    compile 'org.springframework.boot:spring-boot-starter-logging'
-    compile 'org.springframework.boot:spring-boot-starter-tomcat'
-    compile "org.grails:grails-core:$grailsVersion" //The version is mandatory because the dependencies wont be inside the pom
-    compile 'org.grails:grails-dependencies'
-    compile 'org.grails:grails-web-boot'
-    compile 'org.grails.plugins:asset-pipeline:3.1.0'
+subprojects { p ->
+    apply plugin: 'eclipse'
+    apply plugin: 'idea'
+    apply plugin: 'groovy'
 
-    console 'org.grails:grails-console'
+    plugins.withType(JavaPlugin) {
+        sourceCompatibility = 1.8
+    }
 
-    profile 'org.grails.profiles:web-plugin:3.1.3'
+    plugins.withType(GroovyPlugin) {
+        groovydoc {
+            docTitle = "I18n asset pipeline API v${version}"
+            windowTitle = "I18n asset pipeline API v${version}"
+        }
 
-    testCompile 'org.grails:grails-plugin-testing'
-    compile gradleApi()
-    compile localGroovy()
+    }
+
+    repositories {
+        mavenLocal()
+        mavenCentral()
+        maven { url 'https://repo.grails.org/grails/core' }
+    }
+
+    ext {
+        versions = [asset: '2.14.3', groovy: '2.4.11']
+    }
+
+    dependencies {
+        //The version is mandatory because the dependencies wont be inside the pom
+        compile "org.grails:grails-core:$grailsVersion"
+    }
 }
+
 task wrapper(type: Wrapper) {
     gradleVersion = gradleWrapperVersion
-}
-
-groovydoc {
-    docTitle = "I18n asset pipeline API v${version}"
-    windowTitle = "I18n asset pipeline API v${version}"
-}
-
-grailsPublish {
-    user = project.hasProperty('bintrayUser') \
-        ? project.property('bintrayUser')
-        : System.getenv('BINTRAY_USER')
-    key = project.hasProperty('bintrayApiKey') \
-        ? project.property('bintrayApiKey')
-        : System.getenv('BINTRAY_API_KEY')
-    userOrg = 'webbfontaine'
-    githubSlug = 'webbfontaine/i18n-asset-pipeline'
-    license {
-        name = 'Apache-2.0'
-    }
-    title = 'i18n-asset-pipeline'
-    desc = 'asset-pipeline plugin to use localized messages in JavaScript.'
-    developers = [dellermann: 'Daniel Ellermann', anthonymoreira: 'Anthony Moreira']
 }
