+++ i18n-asset-pipeline-core/src/main/groovy/asset/pipeline/i18n/I18nProcessor.groovy
