+++ i18n-asset-pipeline-core/src/main/resources/META-INF/asset-pipeline/asset.specs
