+++ i18n-asset-pipeline-core/src/test/groovy/asset/pipeline/i18n/I18nProcessorSpec.groovy
