+++ i18n-asset-pipeline-core/src/test/resources/assets/texts_de-63afa4054431cd255dd522314d42fe01.js
