+++ i18n-asset-pipeline-gradle/src/main/resources/META-INF/gradle-plugins/org.grails.plugins.i18n-asset-pipeline.properties
@@ -1,0 +1,1 @@
+implementation-class=asset.pipeline.i18n.MessageResearchPlugin
