+++ i18n-asset-pipeline-grails/build.gradle
@@ -1,0 +1,57 @@
+/*
+ * build.gradle
+ *
+ * Copyright (c) 2014-2016, Daniel Ellermann
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+apply plugin: 'org.grails.grails-plugin'
+apply plugin: "org.grails.grails-plugin-publish"
+apply plugin: 'org.grails.grails-gsp'
+apply plugin: 'org.grails.grails-doc'
+
+dependencies {
+    compile 'org.springframework.boot:spring-boot-autoconfigure'
+    compile 'org.springframework.boot:spring-boot-starter-actuator'
+    compile 'org.springframework.boot:spring-boot-starter-logging'
+    compile 'org.springframework.boot:spring-boot-starter-tomcat'
+    compile 'org.grails:grails-dependencies'
+    compile 'org.grails:grails-web-boot'
+    compile "com.bertramlabs.plugins:asset-pipeline-grails:${versions.asset}"
+
+    compile project(':i18n-asset-pipeline-core')
+
+    console 'org.grails:grails-console'
+
+    profile 'org.grails.profiles:web-plugin:3.1.3'
+
+    testCompile 'org.grails:grails-plugin-testing'
+}
+
+grailsPublish {
+    user = project.hasProperty('bintrayUser')    \
+           ? project.property('bintrayUser')
+            : System.getenv('BINTRAY_USER')
+    key = project.hasProperty('bintrayApiKey')    \
+           ? project.property('bintrayApiKey')
+            : System.getenv('BINTRAY_API_KEY')
+    userOrg = 'webbfontaine'
+    githubSlug = "webbfontaine/${project.name}"
+    license {
+        name = 'Apache-2.0'
+    }
+    title = project.name
+    desc = 'asset-pipeline plugin to use localized messages in JavaScript.'
+    developers = [dellermann: 'Daniel Ellermann', anthonymoreira: 'Anthony Moreira', tarikchrouki:"Tarik Chrouki"]
+}
