+++ i18n-asset-pipeline-grails/grails-app/init/i18n/asset/pipeline/Application.groovy
