+++ i18n-asset-pipeline-grails/src/test/resources/assets/texts_de-63afa4054431cd255dd522314d42fe01.js
@@ -1,0 +1,3 @@
+/*
+ * This file is used in unit test I18nTagLibSpec.
+ */
