--- src/main/resources/META-INF/gradle-plugins/org.grails.plugins.i18n-asset-pipeline.properties
@@ -1,1 +1,0 @@
-implementation-class=i18n.MessageResearchPlugin
